Исправил тип координат с int на double

+++ Sandbox/src/main/java/Point.java
@@ -3,7 +3,7 @@ public class Point {
     public double a;
     public double b;
 
-    public Point (int a, int b){
+    public Point (double a, double b){
         this.a=a;
         this.b=b;
     }
